Add an End Scene with Controller
We add a new Scene that can be displayed when the game is over.
We add a new Script EndGameController which runs in Level1 and checks
whether or not the game is over.
If it is, it starts the new Scene.

Currently, the game just ends after ten seconds, since we did
not implement a proper end criterion yet.

Next step: Add a quit and restart button to the end screen.