Removed a self-introduced bug
The  NPCs set their initial condition which was only later on Start() transfered
to the human's condition. Hence the  NPC  list itself did not store  the
right condition (exposed). This was no problem since we did not really do
anything with the list - until now where I tried to use it.