Bugfix for levelstats
The levelstats were not initialized if not started from the main menu
which is a problem if you are not using  it for debugging purposes.
I  put in an "Awake"-function that initializes the  levelstats
before the first  frame if they have not been initialized already
from the main menu.

In addtion, there was a small error in humans base due to merging
that I fixed in this commit.