Added an agent based infection model with COVID-19 params
Since the implementation of SEIR workd but was quite disruptive
I decided to go for a probability based agent based time delayed
implementation which  fits much better to our game (since we
already have the agents). I did a lot of COVID research to
fit different params in the model. So far I also compute the
SEIR equations which are not used. It is one script that works
on its own and could easily be removed in the future. But it
gives a good indication if the agent-based approach  roughly
gives the same results. I plan to do some fine-tuning in the
future.

* Removed abstract infection class: I created the timeDelayed class
similar to the probability based one. However, it seems not feasible to
call the abstract infection class in human base, since the base class
does not know which of its children to call. Even when i automatically
disable one infection model, the base class still tries to call it.
--> if there was another idea behind the abstract infection class I would
have no problem of using it again

* Instead there is now an infection mangager called from the human base
class that decides which infection model to use

* Removed even more infection related apsects from human base into the
infection classes

* in the level settings you can now choose which infection model to choose,
not implemented in menu yet