Create population and friend level 3
Level 3 has its own create population script since all npcs
are created in the center of the dance floor, the player
at the top right edge, the friend at the bottom right edge.
The friend is derived from NPC and will follow the player
once the player collides with him. Both the friend and the player
live in the player-layer which in the project settings  disables
collisions with the edges of the dance floor. The friend only
enters this layer when the player touches him. Both can then move
freely in the scene.