First working version.
The level ends if we bring 3 masks to the hospital.
The hospital needed 2 colliders. One trigger collider
to check for the event that the player enters, but that
does not hinder the player from entering.
And a second collider that only collides with NPCs and keeps
them from entering the hospital.

Currently NPCs may still spawn inside the hospital and cant escape.