Mouseover for endscreen
Added mouse over descriptions for statistics on end screen.
Changed  Killed by you / Infected by you text.
Mouse over script can  now be used by dragging the text in the editor
to the script or by the text being a child of the object that the text
is attached to.