The end game controller base handles call of canvas
The end game controller calls the succeed canvas when
the player is home and healthy and it calls the other one
when the player is exposed. All other end game controllers have
to assign these game objects otherwise the other end game conditions
are taken. For example, the sandbox level still works as usual
since there are no fail/succ  canvas objects to assign. The
end game controller also hanldes the cummulative sprite update.
However, the children still use the CreateHuman-Array-NPC-List
for that which should be done differently in the future.