Added Canvas Fail and Succ to leveldisco
These are now called by the end level controller when the game fails
or is won. Implemnted different fail strategies (player exits without
friend, exits sick without friend, friend is sick, etc.). As soon
as the exit sign is reached the game is over for good or for worse.
The  music only stops when the  player succeeds in getting their friend
out. The different fail possibilities are implemented by changing the
text of the fail screen.