Fixed a bug where we could win and loose at the same time.
The end level controller check for win/loose should now be
mutually exclusive.

Before it could happen that you had all masks, thus triggering the
succ canvas, but then got infected in the background, triggering
the fail canvas.
This should now not be possible anymore.
Once the level is won, it is won.