Update player house
the player house notifies the endlevelcontroller when
the player left.

Also tried to fix a bug where the player can be inside and outside
at the same time.
I made it such that the coroutines for entering and exiting should not be executed simultaneously,
however, the bug is still present.