Endlevelcontroller can end the level only once.
It could have happened that the elc ended the level
because Ãa button was pressed and in the same frame end
it because the playre reached the goal.

We introduced safety measures to prevent this.
A variable levelhasfinished now gets set to true
as soon as the level finishes and the elc does not
check any condition if this var is true.