Make test mode optional for endlevelcontroller
This way, we do not need to remove any code when
we want to switch off test mode.
We just uncheck the test mode box in the scene view
for the endlevel controller.
Also, this allows us to have test mode more some levels
and not for others if we want to.